# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MonkeyCode API",
        "title": "MonkeyCode API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/billing/chat/info": {
            "get": {
                "description": "获取对话内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "获取对话内容",
                "operationId": "chat-info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对话记录ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ChatInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/billing/chat/record": {
            "get": {
                "description": "获取对话记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "获取对话记录",
                "operationId": "list-chat-record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "下一页标识",
                        "name": "next_token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少条记录",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ListChatRecordResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/billing/completion/info": {
            "get": {
                "description": "获取补全内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "获取补全内容",
                "operationId": "completion-info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "补全记录ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.CompletionInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/billing/completion/record": {
            "get": {
                "description": "获取补全记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "获取补全记录",
                "operationId": "list-completion-record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "下一页标识",
                        "name": "next_token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少条记录",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ListCompletionRecordResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/dashboard/user-events": {
            "get": {
                "description": "获取用户事件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "获取用户事件",
                "operationId": "user-events-dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.UserEvent"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/user-heatmap": {
            "get": {
                "description": "用户热力图",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "用户热力图",
                "operationId": "user-heatmap-dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.UserHeatmapResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/user-stat": {
            "get": {
                "description": "获取用户统计信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "获取用户统计信息",
                "operationId": "user-stat-dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.UserStat"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "consts.ModelStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive"
            ],
            "x-enum-varnames": [
                "ModelStatusActive",
                "ModelStatusInactive"
            ]
        },
        "consts.ModelType": {
            "type": "string",
            "enum": [
                "llm",
                "coder",
                "embedding",
                "audio",
                "reranker"
            ],
            "x-enum-varnames": [
                "ModelTypeLLM",
                "ModelTypeCoder",
                "ModelTypeEmbedding",
                "ModelTypeAudio",
                "ModelTypeReranker"
            ]
        },
        "consts.UserStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "locked"
            ],
            "x-enum-varnames": [
                "UserStatusActive",
                "UserStatusInactive",
                "UserStatusLocked"
            ]
        },
...
        "domain.CategoryPoint": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "分类",
                    "type": "string"
                },
                "value": {
                    "description": "值",
                    "type": "integer"
                }
            }
        },
...
        "domain.ChatInfo": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "domain.ChatRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "id": {
                    "description": "记录ID",
                    "type": "string"
                },
                "input_tokens": {
                    "description": "输入token",
                    "type": "integer"
                },
                "model": {
                    "description": "模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Model"
                        }
                    ]
                },
                "output_tokens": {
                    "description": "输出token",
                    "type": "integer"
                },
                "question": {
                    "description": "问题",
                    "type": "string"
                },
                "user": {
                    "description": "用户",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.User"
                        }
                    ]
                },
                "work_mode": {
                    "description": "工作模式",
                    "type": "string"
                }
            }
        },
...
        "domain.CompletionInfo": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "domain.CompletionRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "id": {
                    "description": "记录ID",
                    "type": "string"
                },
                "input_tokens": {
                    "description": "输入token",
                    "type": "integer"
                },
                "is_accept": {
                    "description": "是否采纳",
                    "type": "boolean"
                },
                "output_tokens": {
                    "description": "输出token",
                    "type": "integer"
                },
                "program_language": {
                    "description": "编程语言",
                    "type": "string"
                },
                "user": {
                    "description": "用户",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.User"
                        }
                    ]
                }
            }
        },
...
        "domain.ListChatRecordResp": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "next_token": {
                    "type": "string"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ChatRecord"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "domain.ListCompletionRecordResp": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "next_token": {
                    "type": "string"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.CompletionRecord"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Model": {
            "type": "object",
            "properties": {
                "api_base": {
                    "description": "接口地址 如：https://api.qwen.com",
                    "type": "string"
                },
                "api_key": {
                    "description": "接口密钥 如：sk-xxxx",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "id": {
                    "description": "模型ID",
                    "type": "string"
                },
                "input": {
                    "description": "输入token数",
                    "type": "integer"
                },
                "is_active": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "model_name": {
                    "description": "模型名称 如: deepseek-v3",
                    "type": "string"
                },
                "model_type": {
                    "description": "模型类型 llm:对话模型 coder:代码模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelType"
                        }
                    ]
                },
                "output": {
                    "description": "输出token数",
                    "type": "integer"
                },
                "provider": {
                    "description": "提供商",
                    "type": "string"
                },
                "status": {
                    "description": "状态 active:启用 inactive:禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ModelStatus"
                        }
                    ]
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "integer"
                }
            }
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "last_active_at": {
                    "description": "最后活跃时间",
                    "type": "integer"
                },
                "status": {
                    "description": "用户状态 active: 正常 locked: 锁定 inactive: 禁用",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.UserStatus"
                        }
                    ]
                },
                "two_step_auth": {
                    "description": "是否开启两步验证",
                    "type": "boolean"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "domain.UserEvent": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "事件时间",
                    "type": "integer"
                },
                "name": {
                    "description": "事件名称",
                    "type": "string"
                }
            }
        },
        "domain.UserHeatmap": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "integer"
                }
            }
        },
        "domain.UserHeatmapResp": {
            "type": "object",
            "properties": {
                "max_count": {
                    "type": "integer"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserHeatmap"
                    }
                }
            }
        },
...
        "domain.UserStat": {
            "type": "object",
            "properties": {
                "accepted_per": {
                    "description": "接受率统计",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "timestamp": {
                                "description": "时间戳",
                                "type": "integer"
                            },
                            "value": {
                                "description": "值",
                                "type": "number"
                            }
                        }
                    }
                },
                "chats": {
                    "description": "对话任务数统计",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "timestamp": {
                                "description": "时间戳",
                                "type": "integer"
                            },
                            "value": {
                                "description": "值",
                                "type": "integer"
                            }
                        }
                    }
                },
                "code_completions": {
                    "description": "补全任务数统计",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "timestamp": {
                                "description": "时间戳",
                                "type": "integer"
                            },
                            "value": {
                                "description": "值",
                                "type": "integer"
                            }
                        }
                    }
                },
                "lines_of_code": {
                    "description": "代码行数统计",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "timestamp": {
                                "description": "时间戳",
                                "type": "integer"
                            },
                            "value": {
                                "description": "值",
                                "type": "integer"
                            }
                        }
                    }
                },
                "program_language": {
                    "description": "编程语言占比",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.CategoryPoint"
                    }
                },
                "total_accepted_per": {
                    "description": "近90天总接受率",
                    "type": "number"
                },
                "total_chats": {
                    "description": "近90天总对话任务数",
                    "type": "integer"
                },
                "total_completions": {
                    "description": "近90天总补全任务数",
                    "type": "integer"
                },
                "total_lines_of_code": {
                    "description": "近90天总代码行数",
                    "type": "integer"
                },
                "work_mode": {
                    "description": "工作模式占比",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.CategoryPoint"
                    }
                }
            }
        },
        "web.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}